requirements.txt
amqp==2.6.0
asgiref==3.2.7
Babel==2.8.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bootstrap4==0.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.6
django-bootstrap4==1.1.1
django-cors-headers==3.3.0
django-debug-toolbar==2.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-weasyprint==1.0.1
djangorestframework==3.11.0
easy-thumbnails==2.7
flower==0.9.4
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
humanfriendly==8.2
humanize==0.5.1
idna==2.9
kombu==4.6.10
phonenumbers==8.12.4
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
Pyphen==0.9.5
pyreadline==2.1
pytz==2020.1
redis==3.5.3
requests==2.23.0
scapy==2.4.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
tinycss2==1.0.2
tornado==6.0.4
urllib3==1.25.9
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
wget==3.2
whitenoise==5.1.0
